Reinstall tailwind css
