+++ Procfile.dev
@@ -1,4 +1,3 @@
 web: unset PORT && bin/rails server
 js: yarn build --watch
 css: yarn build:css --watch
-css: bin/rails tailwindcss:watch
