+++ app/assets/stylesheets/application.tailwind.css
@@ -1,7 +1,3 @@
-/*
-require modal
- */
-
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
