+++ package.json
@@ -5,11 +5,14 @@
     "@hotwired/stimulus": "^3.2.1",
     "@hotwired/turbo-rails": "^7.3.0",
     "@rails/request.js": "^0.0.8",
-    "autoprefixer": "^10.4.13",
+    "@tailwindcss/aspect-ratio": "^0.4.2",
+    "@tailwindcss/forms": "^0.5.7",
+    "@tailwindcss/typography": "^0.5.10",
+    "autoprefixer": "^10.4.17",
     "esbuild": "^0.17.10",
     "jquery": "^3.6.4",
-    "postcss": "^8.4.21",
-    "tailwindcss": "^3.2.7",
+    "postcss": "^8.4.35",
+    "tailwindcss": "^3.4.1",
     "yarn": "^1.22.19"
   },
   "scripts": {
